Integrate shadcn/ui charts into dashboard with official examples
