--- app/(logged-in)/charts/page.tsx
@@ -1,586 +1,0 @@
-"use client";
-
-import * as React from "react";
-import { TrendingUp, TrendingDown, Activity, BarChart3, PieChart, LineChart } from "lucide-react";
-import {
-  Area,
-  AreaChart,
-  Bar,
-  BarChart,
-  CartesianGrid,
-  Line,
-  LineChart as RechartsLineChart,
-  Pie,
-  PieChart as RechartsPieChart,
-  RadialBar,
-  RadialBarChart,
-  ResponsiveContainer,
-  XAxis,
-  YAxis,
-  Cell,
-} from "recharts";
-
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
-import {
-  ChartContainer,
-  ChartTooltip,
-  ChartTooltipContent,
-  ChartLegend,
-  ChartLegendContent,
-  type ChartConfig,
-} from "@/components/ui/chart";
-import { Button } from "@/components/ui/button";
-import { Badge } from "@/components/ui/badge";
-import {
-  RadarChartComponent,
-  ComposedChartComponent,
-  PieChartWithCustomShape,
-} from "@/components/charts/advanced-charts";
-
-// Sample data for different chart types
-const monthlyData = [
-  { month: "January", desktop: 186, mobile: 80 },
-  { month: "February", desktop: 305, mobile: 200 },
-  { month: "March", desktop: 237, mobile: 120 },
-  { month: "April", desktop: 73, mobile: 190 },
-  { month: "May", desktop: 209, mobile: 130 },
-  { month: "June", desktop: 214, mobile: 140 },
-];
-
-const trendsData = [
-  { date: "2024-04-01", desktop: 222, mobile: 150 },
-  { date: "2024-04-02", desktop: 97, mobile: 180 },
-  { date: "2024-04-03", desktop: 167, mobile: 120 },
-  { date: "2024-04-04", desktop: 242, mobile: 260 },
-  { date: "2024-04-05", desktop: 373, mobile: 290 },
-  { date: "2024-04-06", desktop: 301, mobile: 340 },
-  { date: "2024-04-07", desktop: 245, mobile: 180 },
-  { date: "2024-04-08", desktop: 409, mobile: 320 },
-  { date: "2024-04-09", desktop: 59, mobile: 110 },
-  { date: "2024-04-10", desktop: 261, mobile: 190 },
-];
-
-const browserData = [
-  { browser: "chrome", visitors: 275, fill: "var(--color-chrome)" },
-  { browser: "safari", visitors: 200, fill: "var(--color-safari)" },
-  { browser: "firefox", visitors: 187, fill: "var(--color-firefox)" },
-  { browser: "edge", visitors: 173, fill: "var(--color-edge)" },
-  { browser: "other", visitors: 90, fill: "var(--color-other)" },
-];
-
-const progressData = [
-  { goal: "target", progress: 73, fill: "var(--color-target)" },
-];
-
-// Chart configurations
-const barChartConfig = {
-  desktop: {
-    label: "Desktop",
-    color: "hsl(var(--chart-1))",
-  },
-  mobile: {
-    label: "Mobile",
-    color: "hsl(var(--chart-2))",
-  },
-} satisfies ChartConfig;
-
-const areaChartConfig = {
-  desktop: {
-    label: "Desktop",
-    color: "hsl(var(--chart-1))",
-  },
-  mobile: {
-    label: "Mobile",
-    color: "hsl(var(--chart-2))",
-  },
-} satisfies ChartConfig;
-
-const lineChartConfig = {
-  desktop: {
-    label: "Desktop",
-    color: "hsl(var(--chart-1))",
-  },
-  mobile: {
-    label: "Mobile",
-    color: "hsl(var(--chart-2))",
-  },
-} satisfies ChartConfig;
-
-const pieChartConfig = {
-  visitors: {
-    label: "Visitors",
-  },
-  chrome: {
-    label: "Chrome",
-    color: "hsl(var(--chart-1))",
-  },
-  safari: {
-    label: "Safari",
-    color: "hsl(var(--chart-2))",
-  },
-  firefox: {
-    label: "Firefox",
-    color: "hsl(var(--chart-3))",
-  },
-  edge: {
-    label: "Edge",
-    color: "hsl(var(--chart-4))",
-  },
-  other: {
-    label: "Other",
-    color: "hsl(var(--chart-5))",
-  },
-} satisfies ChartConfig;
-
-const radialChartConfig = {
-  target: {
-    label: "Target Progress",
-    color: "hsl(var(--chart-2))",
-  },
-} satisfies ChartConfig;
-
-export default function ChartsPage() {
-  const [activeChart, setActiveChart] = React.useState<keyof typeof barChartConfig>("desktop");
-
-  const total = React.useMemo(
-    () => ({
-      desktop: trendsData.reduce((acc, curr) => acc + curr.desktop, 0),
-      mobile: trendsData.reduce((acc, curr) => acc + curr.mobile, 0),
-    }),
-    []
-  );
-
-  const totalVisitors = React.useMemo(() => {
-    return browserData.reduce((acc, curr) => acc + curr.visitors, 0);
-  }, []);
-
-  return (
-    <div className="flex-1 space-y-8 p-8 pt-6">
-      {/* Header */}
-      <div className="flex items-center justify-between space-y-2">
-        <div>
-          <h2 className="text-3xl font-bold tracking-tight">Charts Showcase</h2>
-          <p className="text-muted-foreground">
-            Beautiful charts built with shadcn/ui and Recharts. Copy and paste into your apps.
-          </p>
-        </div>
-        <div className="flex items-center space-x-2">
-          <Badge variant="secondary">shadcn/ui</Badge>
-          <Badge variant="outline">Recharts</Badge>
-        </div>
-      </div>
-
-      {/* Bar Charts Section */}
-      <div className="grid gap-6 md:grid-cols-2">
-        {/* Basic Bar Chart */}
-        <Card>
-          <CardHeader>
-            <CardTitle className="flex items-center gap-2">
-              <BarChart3 className="h-5 w-5" />
-              Bar Chart - Basic
-            </CardTitle>
-            <CardDescription>
-              A basic bar chart showing desktop and mobile users
-            </CardDescription>
-          </CardHeader>
-          <CardContent>
-            <ChartContainer config={barChartConfig}>
-              <BarChart accessibilityLayer data={monthlyData}>
-                <CartesianGrid vertical={false} />
-                <XAxis
-                  dataKey="month"
-                  tickLine={false}
-                  tickMargin={10}
-                  axisLine={false}
-                  tickFormatter={(value) => value.slice(0, 3)}
-                />
-                <ChartTooltip content={<ChartTooltipContent />} />
-                <ChartLegend content={<ChartLegendContent />} />
-                <Bar dataKey="desktop" fill="var(--color-desktop)" radius={4} />
-                <Bar dataKey="mobile" fill="var(--color-mobile)" radius={4} />
-              </BarChart>
-            </ChartContainer>
-          </CardContent>
-          <CardFooter className="flex-col items-start gap-2 text-sm">
-            <div className="flex gap-2 font-medium leading-none">
-              Trending up by 5.2% this month <TrendingUp className="h-4 w-4" />
-            </div>
-            <div className="leading-none text-muted-foreground">
-              January - June 2024
-            </div>
-          </CardFooter>
-        </Card>
-
-        {/* Interactive Bar Chart */}
-        <Card>
-          <CardHeader className="flex flex-col items-stretch space-y-0 border-b p-0 sm:flex-row">
-            <div className="flex flex-1 flex-col justify-center gap-1 px-6 py-5 sm:py-6">
-              <CardTitle className="flex items-center gap-2">
-                <Activity className="h-5 w-5" />
-                Bar Chart - Interactive
-              </CardTitle>
-              <CardDescription>
-                Showing total visitors for the last 3 months
-              </CardDescription>
-            </div>
-            <div className="flex">
-              {["desktop", "mobile"].map((key) => {
-                const chart = key as keyof typeof barChartConfig;
-                return (
-                  <button
-                    key={chart}
-                    data-active={activeChart === chart}
-                    className="relative z-30 flex flex-1 flex-col justify-center gap-1 border-t px-6 py-4 text-left even:border-l data-[active=true]:bg-muted/50 sm:border-l sm:border-t-0 sm:px-8 sm:py-6"
-                    onClick={() => setActiveChart(chart)}
-                  >
-                    <span className="text-xs text-muted-foreground">
-                      {barChartConfig[chart].label}
-                    </span>
-                    <span className="text-lg font-bold leading-none sm:text-3xl">
-                      {total[key as keyof typeof total].toLocaleString()}
-                    </span>
-                  </button>
-                );
-              })}
-            </div>
-          </CardHeader>
-          <CardContent className="px-2 sm:p-6">
-            <ChartContainer
-              config={barChartConfig}
-              className="aspect-auto h-[250px] w-full"
-            >
-              <BarChart
-                accessibilityLayer
-                data={trendsData}
-                margin={{
-                  left: 12,
-                  right: 12,
-                }}
-              >
-                <CartesianGrid vertical={false} />
-                <XAxis
-                  dataKey="date"
-                  tickLine={false}
-                  axisLine={false}
-                  tickMargin={8}
-                  minTickGap={32}
-                  tickFormatter={(value) => {
-                    const date = new Date(value);
-                    return date.toLocaleDateString("en-US", {
-                      month: "short",
-                      day: "numeric",
-                    });
-                  }}
-                />
-                <ChartTooltip
-                  content={
-                    <ChartTooltipContent
-                      className="w-[150px]"
-                      nameKey="views"
-                      labelFormatter={(value) => {
-                        return new Date(value).toLocaleDateString("en-US", {
-                          month: "short",
-                          day: "numeric",
-                          year: "numeric",
-                        });
-                      }}
-                    />
-                  }
-                />
-                <Bar dataKey={activeChart} fill={`var(--color-${activeChart})`} />
-              </BarChart>
-            </ChartContainer>
-          </CardContent>
-        </Card>
-      </div>
-
-      {/* Area and Line Charts */}
-      <div className="grid gap-6 md:grid-cols-2">
-        {/* Area Chart */}
-        <Card>
-          <CardHeader>
-            <CardTitle>Area Chart</CardTitle>
-            <CardDescription>
-              Showing total visitors for the last 6 months
-            </CardDescription>
-          </CardHeader>
-          <CardContent>
-            <ChartContainer config={areaChartConfig}>
-              <AreaChart
-                accessibilityLayer
-                data={monthlyData}
-                margin={{
-                  left: 12,
-                  right: 12,
-                }}
-              >
-                <CartesianGrid vertical={false} />
-                <XAxis
-                  dataKey="month"
-                  tickLine={false}
-                  axisLine={false}
-                  tickMargin={8}
-                  tickFormatter={(value) => value.slice(0, 3)}
-                />
-                <ChartTooltip content={<ChartTooltipContent />} />
-                <Area
-                  dataKey="mobile"
-                  type="natural"
-                  fill="var(--color-mobile)"
-                  fillOpacity={0.4}
-                  stroke="var(--color-mobile)"
-                  stackId="a"
-                />
-                <Area
-                  dataKey="desktop"
-                  type="natural"
-                  fill="var(--color-desktop)"
-                  fillOpacity={0.4}
-                  stroke="var(--color-desktop)"
-                  stackId="a"
-                />
-              </AreaChart>
-            </ChartContainer>
-          </CardContent>
-          <CardFooter>
-            <div className="flex w-full items-start gap-2 text-sm">
-              <div className="grid gap-2">
-                <div className="flex items-center gap-2 font-medium leading-none">
-                  Trending up by 5.2% this month <TrendingUp className="h-4 w-4" />
-                </div>
-                <div className="flex items-center gap-2 leading-none text-muted-foreground">
-                  January - June 2024
-                </div>
-              </div>
-            </div>
-          </CardFooter>
-        </Card>
-
-        {/* Line Chart */}
-        <Card>
-          <CardHeader>
-            <CardTitle className="flex items-center gap-2">
-              <LineChart className="h-5 w-5" />
-              Line Chart
-            </CardTitle>
-            <CardDescription>January - June 2024</CardDescription>
-          </CardHeader>
-          <CardContent>
-            <ChartContainer config={lineChartConfig}>
-              <RechartsLineChart
-                accessibilityLayer
-                data={monthlyData}
-                margin={{
-                  left: 12,
-                  right: 12,
-                }}
-              >
-                <CartesianGrid vertical={false} />
-                <XAxis
-                  dataKey="month"
-                  tickLine={false}
-                  axisLine={false}
-                  tickMargin={8}
-                  tickFormatter={(value) => value.slice(0, 3)}
-                />
-                <ChartTooltip content={<ChartTooltipContent />} />
-                <Line
-                  dataKey="desktop"
-                  type="monotone"
-                  stroke="var(--color-desktop)"
-                  strokeWidth={2}
-                  dot={false}
-                />
-                <Line
-                  dataKey="mobile"
-                  type="monotone"
-                  stroke="var(--color-mobile)"
-                  strokeWidth={2}
-                  dot={false}
-                />
-              </RechartsLineChart>
-            </ChartContainer>
-          </CardContent>
-          <CardFooter>
-            <div className="flex w-full items-start gap-2 text-sm">
-              <div className="grid gap-2">
-                <div className="flex items-center gap-2 font-medium leading-none">
-                  Trending up by 5.2% this month <TrendingUp className="h-4 w-4" />
-                </div>
-                <div className="flex items-center gap-2 leading-none text-muted-foreground">
-                  Showing total visitors for the last 6 months
-                </div>
-              </div>
-            </div>
-          </CardFooter>
-        </Card>
-      </div>
-
-      {/* Pie and Radial Charts */}
-      <div className="grid gap-6 md:grid-cols-2">
-        {/* Pie Chart */}
-        <Card className="flex flex-col">
-          <CardHeader className="items-center pb-0">
-            <CardTitle className="flex items-center gap-2">
-              <PieChart className="h-5 w-5" />
-              Pie Chart - Donut
-            </CardTitle>
-            <CardDescription>January - June 2024</CardDescription>
-          </CardHeader>
-          <CardContent className="flex-1 pb-0">
-            <ChartContainer
-              config={pieChartConfig}
-              className="mx-auto aspect-square max-h-[250px]"
-            >
-              <RechartsPieChart>
-                <ChartTooltip
-                  cursor={false}
-                  content={<ChartTooltipContent hideLabel />}
-                />
-                <Pie
-                  data={browserData}
-                  dataKey="visitors"
-                  nameKey="browser"
-                  innerRadius={60}
-                  strokeWidth={5}
-                >
-                  <Cell fill="var(--color-chrome)" />
-                  <Cell fill="var(--color-safari)" />
-                  <Cell fill="var(--color-firefox)" />
-                  <Cell fill="var(--color-edge)" />
-                  <Cell fill="var(--color-other)" />
-                </Pie>
-              </RechartsPieChart>
-            </ChartContainer>
-          </CardContent>
-          <CardFooter className="flex-col gap-2 text-sm">
-            <div className="flex items-center gap-2 font-medium leading-none">
-              Trending up by 5.2% this month <TrendingUp className="h-4 w-4" />
-            </div>
-            <div className="leading-none text-muted-foreground">
-              Showing total visitors for the last 6 months
-            </div>
-          </CardFooter>
-        </Card>
-
-        {/* Radial Chart */}
-        <Card className="flex flex-col">
-          <CardHeader className="items-center pb-2">
-            <CardTitle>Radial Chart - Shape</CardTitle>
-            <CardDescription>Goal Progress</CardDescription>
-          </CardHeader>
-          <CardContent className="flex-1 pb-0">
-            <ChartContainer
-              config={radialChartConfig}
-              className="mx-auto aspect-square max-h-[250px]"
-            >
-              <RadialBarChart
-                data={progressData}
-                startAngle={0}
-                endAngle={250}
-                innerRadius={80}
-                outerRadius={110}
-              >
-                <ChartTooltip
-                  cursor={false}
-                  content={<ChartTooltipContent hideLabel nameKey="goal" />}
-                />
-                <RadialBar dataKey="progress" background cornerRadius={10} />
-              </RadialBarChart>
-            </ChartContainer>
-          </CardContent>
-          <CardFooter className="flex-col gap-2 text-sm">
-            <div className="flex items-center gap-2 font-medium leading-none">
-              Goal completion at 73% <TrendingUp className="h-4 w-4" />
-            </div>
-            <div className="leading-none text-muted-foreground">
-              Target achievement progress
-            </div>
-          </CardFooter>
-        </Card>
-      </div>
-
-      {/* Advanced Charts Section */}
-      <div className="space-y-6">
-        <div>
-          <h3 className="text-2xl font-bold tracking-tight">Advanced Charts</h3>
-          <p className="text-muted-foreground">
-            More complex chart types for specialized data visualization needs.
-          </p>
-        </div>
-        
-        <div className="grid gap-6 md:grid-cols-2 lg:grid-cols-3">
-          <RadarChartComponent />
-          <ComposedChartComponent />
-          <PieChartWithCustomShape />
-        </div>
-      </div>
-
-      {/* Documentation Links */}
-      <Card>
-        <CardHeader>
-          <CardTitle>Documentation & Resources</CardTitle>
-          <CardDescription>
-            Learn more about implementing charts with shadcn/ui
-          </CardDescription>
-        </CardHeader>
-        <CardContent>
-          <div className="grid gap-4 md:grid-cols-3">
-            <div className="space-y-2">
-              <h4 className="font-medium">shadcn/ui Charts</h4>
-              <p className="text-sm text-muted-foreground">
-                Official documentation for chart components
-              </p>
-              <Button variant="outline" size="sm" asChild>
-                <a
-                  href="https://ui.shadcn.com/docs/components/chart"
-                  target="_blank"
-                  rel="noopener noreferrer"
-                >
-                  View Docs
-                </a>
-              </Button>
-            </div>
-            <div className="space-y-2">
-              <h4 className="font-medium">Recharts Library</h4>
-              <p className="text-sm text-muted-foreground">
-                The underlying charting library used by shadcn/ui
-              </p>
-              <Button variant="outline" size="sm" asChild>
-                <a
-                  href="https://recharts.org/en-US/"
-                  target="_blank"
-                  rel="noopener noreferrer"
-                >
-                  View Library
-                </a>
-              </Button>
-            </div>
-            <div className="space-y-2">
-              <h4 className="font-medium">Chart Examples</h4>
-              <p className="text-sm text-muted-foreground">
-                More chart examples and implementations
-              </p>
-              <Button variant="outline" size="sm" asChild>
-                <a
-                  href="https://ui.shadcn.com/charts"
-                  target="_blank"
-                  rel="noopener noreferrer"
-                >
-                  Browse Charts
-                </a>
-              </Button>
-            </div>
-          </div>
-        </CardContent>
-      </Card>
-    </div>
-  );
-}
