+++ components/app-sidebar.tsx
@@ -12,7 +12,6 @@ import {
   Send,
   Settings2,
   SquareTerminal,
-  BarChart3,
 } from 'lucide-react';
 import Link from 'next/link';
 
@@ -37,33 +36,6 @@ const data = {
       url: '/dashboard',
       icon: SquareTerminal,
       isActive: true,
-    },
-    {
-      title: 'Charts',
-      url: '/charts',
-      icon: BarChart3,
-      items: [
-        {
-          title: 'All Charts',
-          url: '/charts',
-        },
-        {
-          title: 'Bar Charts',
-          url: '/charts#bar-charts',
-        },
-        {
-          title: 'Line Charts',
-          url: '/charts#line-charts',
-        },
-        {
-          title: 'Pie Charts',
-          url: '/charts#pie-charts',
-        },
-        {
-          title: 'Advanced',
-          url: '/charts#advanced-charts',
-        },
-      ],
     },
     {
       title: 'Playground',
